+++ VarState.h
@@ -22,6 +22,6 @@ struct VarState {
 	void initEMatrix();
 	friend VarState operator +(VarState a, VarState b);
 	friend VarState operator -(VarState a, VarState b);
-	friend VarState operator *(float a, VarState b);
+	friend VarState operator *(double a, VarState b);
 	VarState& operator=(const VarState& other);
 };
